おすすめ機能に関する操作をRecommendManager.javaに集約
