--- WEB-INF/src/controller/RecommendGetManager.java
@@ -1,5 +1,0 @@
-package controller;
-
-public class RecommendGetManager {
-
-}
